pyproject.toml
[project]
name = "start-thkinking-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "openai>=2.8.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
]
